+++ x/fungible/keeper/deposits_test.go
@@ -1,9 +1,10 @@
 package keeper_test
 
 import (
-	"github.com/zeta-chain/node/e2e/contracts/testabort"
 	"math/big"
 	"testing"
+
+	"github.com/zeta-chain/node/e2e/contracts/testabort"
 
 	"cosmossdk.io/math"
 	"github.com/pkg/errors"
@@ -492,7 +493,7 @@ func TestKeeper_ZRC20DepositAndCallContract(t *testing.T) {
 		require.EqualValues(t, int64(0), balance.Int64())
 	})
 
-	t.Run("can deposit using V2", func(t *testing.T) {
+	t.Run("can deposit GAS token using V2", func(t *testing.T) {
 		// setup gas coin
 		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
 		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
@@ -524,6 +525,38 @@ func TestKeeper_ZRC20DepositAndCallContract(t *testing.T) {
 		balance, err := k.BalanceOfZRC4(ctx, zrc20, to)
 		require.NoError(t, err)
 		require.Equal(t, big.NewInt(42), balance)
+	})
+
+	t.Run("can deposit ZETA token using V2", func(t *testing.T) {
+		// setup gas coin
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainList := chains.DefaultChainsList()
+		chain := chainList[0].ChainId
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		_ = setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chain, "foobar", "foobar")
+
+		// deposit
+		to := sample.EthAddress()
+		_, contractCall, err := k.ZRC20DepositAndCallContract(
+			ctx,
+			sample.EthAddress().Bytes(),
+			to,
+			big.NewInt(42),
+			chain,
+			[]byte{},
+			coin.CoinType_Zeta,
+			sample.EthAddress().String(),
+			crosschaintypes.ProtocolContractVersion_V2,
+			false,
+		)
+		require.NoError(t, err)
+		require.False(t, contractCall)
+		require.Equal(t, big.NewInt(42).Int64(),
+			sdkk.BankKeeper.GetBalance(ctx, to.Bytes(), config.BaseDenom).Amount.Int64())
 	})
 }
 
@@ -565,6 +598,34 @@ func TestKeeper_DepositCoinZeta(t *testing.T) {
 }
 
 func TestKeeper_ProcessDeposit(t *testing.T) {
+	t.Run("should not process deposit of unsupported coin-type", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+		receiver := sample.EthAddress()
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		zrc20 := setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+
+		// ACT
+		_, _, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			zrc20,
+			receiver,
+			big.NewInt(42),
+			[]byte{},
+			coin.CoinType_Cmd,
+			false,
+		)
+
+		// ASSERT
+		require.ErrorIs(t, err, types.ErrProcessDeposit)
+	})
 	t.Run("should process no-call deposit", func(t *testing.T) {
 		// ARRANGE
 		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
@@ -597,6 +658,38 @@ func TestKeeper_ProcessDeposit(t *testing.T) {
 		balance, err := k.BalanceOfZRC4(ctx, zrc20, receiver)
 		require.NoError(t, err)
 		require.Equal(t, big.NewInt(42), balance)
+	})
+
+	t.Run("should process no-call deposit for coinType ZETA", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+		receiver := sample.EthAddress()
+		amount := big.NewInt(42)
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		_ = setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+
+		// ACT
+		_, contractCall, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			common.Address{}, // empty address for ZETA
+			receiver,
+			amount,
+			[]byte{},
+			coin.CoinType_Zeta,
+			false,
+		)
+
+		// ASSERT
+		require.NoError(t, err)
+		require.False(t, contractCall)
+		require.Equal(t, sdkk.BankKeeper.GetBalance(ctx, receiver.Bytes(), config.BaseDenom).Amount.Int64(), amount.Int64())
 	})
 
 	t.Run("should process no-call deposit, message should be ignored", func(t *testing.T) {
@@ -667,6 +760,141 @@ func TestKeeper_ProcessDeposit(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, big.NewInt(82), balance)
 		assertTestDAppV2MessageAndAmount(t, ctx, k, testDapp, "foo", 82)
+	})
+
+	t.Run("should process no asset call", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+
+		// deploy test dapp
+		testDapp := deployTestDAppV2(t, ctx, k, sdkk.EvmKeeper)
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		zrc20 := setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+
+		// ACT
+		_, contractCall, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			zrc20,
+			testDapp,
+			big.NewInt(82),
+			[]byte("foo"),
+			coin.CoinType_NoAssetCall,
+			false,
+		)
+
+		// ASSERT
+		require.NoError(t, err)
+		require.True(t, contractCall)
+		balance, err := k.BalanceOfZRC4(ctx, zrc20, testDapp)
+		require.NoError(t, err)
+		require.Equal(t, int64(0), balance.Int64())
+		assertTestDAppV2MessageAndAmount(t, ctx, k, testDapp, "foo", 82)
+	})
+
+	t.Run("should process deposit and call for coinType Zeta", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+
+		// deploy test dapp
+		testDapp := deployTestDAppV2(t, ctx, k, sdkk.EvmKeeper)
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		_ = setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+		amount := big.NewInt(82)
+
+		// ACT
+		_, contractCall, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			common.Address{},
+			testDapp,
+			amount,
+			[]byte("foo"),
+			coin.CoinType_Zeta,
+			true,
+		)
+
+		// ASSERT
+		require.NoError(t, err)
+		require.True(t, contractCall)
+		require.Equal(t, sdkk.BankKeeper.GetBalance(ctx, testDapp.Bytes(), config.BaseDenom).Amount.Int64(), amount.Int64())
+		assertTestDAppV2MessageAndAmount(t, ctx, k, testDapp, "foo", 82)
+	})
+
+	t.Run("should process deposit for coinType Zeta", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+
+		// deploy test dapp
+		receiverAdddress := sample.EthAddress()
+
+		// deploy the system contracts
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		_ = setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+		amount := big.NewInt(82)
+
+		// ACT
+		_, contractCall, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			common.Address{},
+			receiverAdddress,
+			amount,
+			[]byte("foo"),
+			coin.CoinType_Zeta,
+			false,
+		)
+
+		// ASSERT
+		require.NoError(t, err)
+		require.False(t, contractCall)
+		require.Equal(t, sdkk.BankKeeper.GetBalance(ctx, receiverAdddress.Bytes(), config.BaseDenom).Amount.Int64(), amount.Int64())
+	})
+
+	t.Run("should not process deposit for coinType Zeta if system contract is not present", func(t *testing.T) {
+		// ARRANGE
+		k, ctx, _, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := chains.DefaultChainsList()[0].ChainId
+
+		// deploy test dapp
+		receiverAdddress := sample.EthAddress()
+
+		amount := big.NewInt(82)
+
+		// ACT
+		_, contractCall, err := k.ProcessDeposit(
+			ctx,
+			sample.EthAddress().Bytes(),
+			chainID,
+			common.Address{},
+			receiverAdddress,
+			amount,
+			[]byte("foo"),
+			coin.CoinType_Zeta,
+			false,
+		)
+
+		// ASSERT
+		require.False(t, contractCall)
+		require.ErrorIs(t, err, types.ErrSystemContractNotFound)
 	})
 
 	t.Run("should process deposit and call with no message", func(t *testing.T) {
@@ -790,7 +1018,7 @@ func TestKeeper_ProcessAbort(t *testing.T) {
 		require.Equal(t, big.NewInt(82), balance)
 	})
 
-	t.Run("can't process abort for ZETA token", func(t *testing.T) {
+	t.Run("successfully process abort for ZETA token", func(t *testing.T) {
 		// ARRANGE
 		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
 		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
@@ -799,6 +1027,7 @@ func TestKeeper_ProcessAbort(t *testing.T) {
 
 		// deploy test dapp
 		testAbort := deployTestAbort(t, ctx, k, sdkk.EvmKeeper)
+		abortAmount := big.NewInt(82)
 
 		// deploy the system contracts
 		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
@@ -807,7 +1036,7 @@ func TestKeeper_ProcessAbort(t *testing.T) {
 		_, err := k.ProcessAbort(
 			ctx,
 			sample.EthAddress().String(),
-			big.NewInt(82),
+			abortAmount,
 			false,
 			chainID,
 			coin.CoinType_Zeta,
@@ -817,8 +1046,9 @@ func TestKeeper_ProcessAbort(t *testing.T) {
 		)
 
 		// ASSERT
-		require.Error(t, err)
-		require.NotErrorIs(t, err, types.ErrOnAbortFailed)
+		require.NoError(t, err)
+		require.Equal(t, abortAmount.Int64(),
+			sdkk.BankKeeper.GetBalance(ctx, testAbort.Bytes(), config.BaseDenom).Amount.Int64())
 	})
 
 	t.Run("can't process abort for invalid chain ID", func(t *testing.T) {