+++ zetaclient/chains/evm/common/constant.go
@@ -71,4 +71,7 @@ const (
 	// TopicsERC20CustodyWithdrawAndCall is the number of topics for an ERC20 custody withdraw and call event
 	// [signature, recipient, asset]
 	TopicsERC20CustodyWithdrawAndCall = 3
+
+	// TopicsZetaConnectorWithdraw is the number of topics for a Zeta connector withdraw event
+	TopicsZetaConnectorWithdraw = 2 // [to, amount]
 )