+++ zetaclient/zetacore/client_test.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"net"
 	"testing"
+	"time"
 
 	sdkmath "cosmossdk.io/math"
-
+	coretypes "github.com/cometbft/cometbft/rpc/core/types"
+	cometbfttypes "github.com/cometbft/cometbft/types"
 	"github.com/cosmos/cosmos-sdk/testutil/mock"
 	"github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
@@ -23,8 +25,6 @@ import (
 	"go.uber.org/mock/gomock"
 
 	cometbftrpc "github.com/cometbft/cometbft/rpc/client"
-	coretypes "github.com/cometbft/cometbft/rpc/core/types"
-	cometbfttypes "github.com/cometbft/cometbft/types"
 	"github.com/zeta-chain/node/cmd/zetacored/config"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/keys"
@@ -249,21 +249,25 @@ func TestZetacore_SubscribeNewBlocks(t *testing.T) {
 		withCometBFT(cometBFTClient),
 	)
 
-	newBlockChan, err := client.NewBlockSubscriber(ctx)
-	require.NoError(t, err)
-
-	height := int64(10)
-
-	cometBFTClient.PublishToSubscribers(coretypes.ResultEvent{
+	expectedHeight := int64(10)
+	blockEvent := coretypes.ResultEvent{
 		Data: cometbfttypes.EventDataNewBlock{
 			Block: &cometbfttypes.Block{
 				Header: cometbfttypes.Header{
-					Height: height,
+					Height: expectedHeight,
 				},
 			},
 		},
-	})
+	}
 
-	newBlockEvent := <-newBlockChan
-	require.Equal(t, height, newBlockEvent.Block.Header.Height)
+	newBlockChan, err := client.NewBlockSubscriber(ctx)
+	require.NoError(t, err)
+
+	go func() {
+		time.Sleep(10 * time.Millisecond)
+		cometBFTClient.PublishToSubscribers(blockEvent)
+	}()
+
+	receivedBlock := <-newBlockChan
+	require.Equal(t, expectedHeight, receivedBlock.Block.Header.Height)
 }